Update LoxInterpreter, GenerateAst, LoxInterpreter.Tests
Refactored and updated GenerateAst to nest children classes as
subclasses inside of the base class. Added a new super class with
children class for chapter 8.1.

Updated the interpreter and the parser in LoxInterpreter to reflect
changes in nesting the subclasses of the Expr superclass, and the
addition of the Stmt superclass. The functionality introduced in 8.1
adds the print function in Lox, and print and an expression-ending
semicolon is now required to display the results of evaluating an expression. This new functionality resolves issue #16.

Updated LoxInterpreter.Tests to terminate expressions with semicolons,
and added casting to certain Interpreter and Parser functions.

Updated README.md with instructions for using dotnet build with only one
project.